requirements.txt
gradio
python-dotenv
openai
requests